[CHANGED] int cast

+++ Assets/Content/Scripts/CardController.cs
@@ -392,7 +392,7 @@ namespace Content.Scripts
             {
                 if (step < 0)
                 {
-                    step = Random.Range(4, 12);
+                    step = (int)Random.Range(4, 12);
                 }
                 this.cardHand.Insert(step, this.cardsById[id]);
                 Debug.LogFormat("Follow Up Card with ID {0} inserted in {1} cards.", id, step + 1);
